% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mining.R
\name{break_str}
\alias{break_str}
\title{Breaking a string into rows of strings}
\usage{
break_str(string, nchar)
}
\arguments{
\item{string}{: a string with many characters}

\item{nchar}{: number of characters per row}
}
\value{
rows of strings
}
\description{
Breaking a string into rows of strings
}
