% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{extract_abstract_pubmed}
\alias{extract_abstract_pubmed}
\title{Extracting the abstract from pubmed}
\usage{
extract_abstract_pubmed(link)
}
\arguments{
\item{link}{: link obtained by the function extract_searchpage_pubmed()}
}
\value{
abstract of the article (a string)
}
\description{
Extracting the abstract from pubmed
}
\examples{
c(title, link) \%<-\% extract_searchpage_pubmed(words2link_pubmed("ptsd ketamine"), 211)
abstract = c()
for (i in 1:10) {
  abstract = c(abstract, extract_abstract_pubmed(link[i]))
}
}
