% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{extract_searchpage_pubmed}
\alias{extract_searchpage_pubmed}
\title{Extracting the searchpage in pubmed}
\usage{
extract_searchpage_pubmed(path, no_of_results)
}
\arguments{
\item{path}{: path generated by the function words2link_pubmed()}

\item{no_of_results}{: total results appeared in pubmed}
}
\value{
two vectors of variables: titles and hyperlinks
}
\description{
Extracting the searchpage in pubmed
}
\examples{
library(zeallot)
c(title, link) \%<-\% extract_searchpage_pubmed(link, 10) # the first 10 entries
}
