% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mining.R
\name{strings2dataframe}
\alias{strings2dataframe}
\title{Breaking strings into rows of strings}
\usage{
strings2dataframe(strings, nchar)
}
\arguments{
\item{nchar}{: number of characters per row}

\item{string}{: strings, each with many character (could be output of functions like extract_abstract_pubmed())}
}
\value{
a reading-friendly dataframe
}
\description{
Breaking strings into rows of strings
}
