% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{words2link_pubmed}
\alias{words2link_pubmed}
\title{Converting keywords to the link of the search result page in pubmed}
\usage{
words2link_pubmed(words)
}
\arguments{
\item{words}{: pubmed search terms}
}
\value{
the link of the search result page in pubmed (size == 200)
}
\description{
Converting keywords to the link of the search result page in pubmed
}
\examples{
words2link_pubmed("(PTSD OR Depression) AND Ketamine")
}
